Complete implementation of search term validation function for Staff Plan Lookup Utility
Once pre-processing is complete, split the input into an array on space seperations, using the quote watcher object move through the array and combine array elements that make up a single search term (ex. "this is a single search term") while also accepting single element non-quoted inputs, check for mismatched quotes, stop the search, and alert the user if detected, and store the verfied input in a hidden input as a list of CSVs to be passed using GET to search.php (this set up keeps the textbox visible to the user from changing to final formatted CSV string).